Use more descriptive variable names
One downside of replacing

   val c = Counter.new 0

   ... Counter.next c ...

with

   val c = Counter.generator 0

   ... c () ...

is that former is more clear that a counter is being incremented.

When the `val c = Counter.generator 0` and the `c ()` are relatively
close together, then it is may not be difficult to understand the
meaning of the `c ()`.  But, when they are far apart, it can be
difficult to understand the meaning of `c ()`.  One way to mitigate is
to choose better variable names:

  val nextThing = Counter.generator 0

  ... nextThing () ...