Experiment using LLVM 'cc10' (aka, 'ghccc')
The 'cc10' (aka, 'ghccc') calling convention has no callee-save
registers.  For chunks that want to use many registers, this avoids
the save and restore of callee-save registers on entry and exit of the
chunk.  A simple "shim" function coordinates between the 'ccc' calling
convention and the 'cc10' calling convention for the dispatch from
`c-main.h`; with `-chunk-tail-call true`, this is the only time that a
shim function is called.