Machine.Operand.isLocation -> isDestination, fix odd backend behaviour
Backend doesn't generate Binds for destinations that were turned into
constants. As it happens, idempotency for existing constants turned
any such binds to self-moves, which although were ill-typed
(e.g. it could in theory include a move from the constant 3 into itself),
were eliminated to Noops by the move function

isLocation conflated two meanings: the ability to point to something,
and the ability to be a destination for moves.
Statics are constant pointers, and such are only in the latter class.
The former is not needed, and can be recovered from the type