Don't mark any functions as `noreturn`
The only `noreturn` functions are `MLton_halt` and `MLton_bug`, which
are only called at program termination, so there is no performance
advantage to marking these functions as `noreturn`.

As noted in a03514280, LLVM will erroneously optimize a `musttail
call` to a `noreturn` function.

For C calls that do not return, we simply leave the chunk with `return
-2;`, much as `Thread_returnToC` does.