Use C99/C11 `inline` for primitive and Basis Library functions
Previously, functions meant to be inlined because they correspond to
primitives or Basis Library functions (e.g., `Real<N>` and
`Real<N>.Math`) were marked `static inline` when included via
`c-chunk.h`.  If a C compiler chooses not to inline a function (such
as clang at -O1), then each .o file included its own copy of the
function (and the copy of the function provided by `libmlton.a` was
not linked into the final executable).

Now, functions are marked `inline` when included via `c-chunk.h` (and
the corresponding `_import` is given the `inline` attribute).  The
C99/C11 semantics of `inline` requires the C compiler to *not* include
a copy of the function in the .o file (if it chooses not to inline the
function) and treat the function as an external reference.  The copy
of the function provided by `libmlton.a` is used to satisfy the
external reference when linking.