Replace isMutable in Rssa binds with pinned
A pinned variable in Rssa is created by splitting live ranges
in bounce-vars, and so the only difference is that it will not
be copy-propagated as a variable (but may be once it is allocated
to a location in backend). The alternative would require that copy
propagation calculate complicated liveness information every time
to determine if a variable can be safely merged with another.