Don't declare `Heap` statics in `.c` and `.ll` files defining `Chunk<N>` fns
`Heap` statics are declared `const static` in the "main" `.c` file and will not
be linkable from the `.c` and `.ll` files defining `Chunk<N>` functions.

In practice, there won't be any references to `Heap` statics from a `Chunk<N>`
function (they will be referenced as global objptrs into the heap) and the
symbol won't appear in the to-be-linked object file.  Nonetheless, it is easy to
skip emitting declarations for `Heap` statics that are known not to be
referenced.