Change `metadata: WordXVector.t` field of `Static.T` to `metadata: WordX.t list`
The metadata of an object need not have all fields of the same size.
In particular, a sequence object's metadata is comprised of a counter
field (of `WordSize.seqIndex` size), a length field (also of
`WordSize.seqIndex` size), and a header.  We happen to use 64-bit
sequence indices and 64-bit headers on 64-bit platforms, but that is
not required.  (Albeit, it is also not tested.)