Ensure that `sizeofInitialBytesLive` returns a non-zero size
With b3873f1d8, a program can have an empty `objectInits[]` to be
allocated in the initial dynamic heap.  With an empty `objectInits[]`,
`sizeofInitialBytesLive` would return 0.  Attempting to create a heap
of zero bytes would fail (the specified SUSv3 behavior of `mmap`),
triggering a backoff computed as:

    highSize = newSize - s->sysvals.pageSize;
    newSize = align((factor-1) * (highSize / factor) + (lowSize / factor), s->sysvals.pageSize);

However, `newSize - s->sysval.pageSize` (with `newSize` equals 0),
wraps to nearly 2^64.  Successive backoffs eventually bring the
requested size down to one that can be satisfied, but the size is
generally much larger than required.  Moreover, the heap won't be
resized until a subsequent GC, which may not occur during the run of
the program, due to the heap being so large.  Many regression tests
run, but most that fork fail with `unhandled exception: SysErr: Cannot
allocate memory [nomem]`, because both the parent has allocated a very
large heap and the child is requesting to allocate a very large heap.

Now, `sizeofInitialBytesLive` also includes the size of the initial
thread/stack to be created, ensuring that it is always non-zero.