Update `XmlTree.Exp.enterLeave` to avoid handler for non-raising
The `XmlTee.Exp.enterLeave` function is used to add profiling
information to XML expressions; it transforms an expression of the
form:

    let
      ... decs ...
    in
      res
    end

to

    let
      val z =
        let
          Enter
          ... decs ...
          Leave
        in
          res
        end
        handle exn => let
                        Leave
                        val x = raise exn
                      in
                        x
                      end
     in
       z
     end

The `handler` wrapper is unnecessary when the original expression
cannot raise; in particular, for a curried function, the body of the
outer function does not need a `handler` wrapper.  This avoids
unnecessarily perturbing the program (relative to the program without
profiling).  For an expression that does not raise, it now simply
transforms to:

    let
      Enter
      ... decs ...
      Leave
    in
      res
    end