Handle jumps to start block in `Ssa{,2}.Profile.addProfileFunction`
If an SSA{,2} function has an internal jump to the start block, then
adding an `Enter <f>` profile statement to the start block will result
in mis-matched sources.  When entering the function, the source list
for the start block will be empty, but when jumping to the start block
from within the function, the source list will be the singleton
`[<f>]`.

When an SSA{,2} function has an internal jump to the start block, then
create a new start block for the `Enter <f>` profile statement.