Disable `Zone` pass for bootstrap
See MLton/mlton#334.

When self-compiling with `Zone` enabled, the resulting compiler (when
self-compiling) has over 2GB live data at the end of `amd64Codegen`.
However, the should be very little live data at the end of
`amd64Codegen`, because the IR program is no longer live.  The `Zone`
pass is known to not be safe-for-space.

The `Zone` pass only transforms SSA2 functions with over 10000 blocks.
MLton's `main` SSA2 function only recently crossed the threshold that
results in the `Zone` pass transforming the `main` SSA2 function,
which is why this space leak only recently became a problem.