Share `Property` definitions for variables in `BounceVars`
See MLton/mlton#334.

This was meant to addresses an issue where the during `bounceVars`,
the `property list maxLength` jumped from 4 to 761 during `bounceVars`
on a self-compile.  (Although, it didn't seem to negatively effect
property list accesses, since the `property list average position`
only jumped from 0.148 to 0.172.)  However, it does not actually
reduce the `property list maxLength` jump, which seems to be coming
from a different pass.