Reset table for hash-consing of {{,S}XML,SSA{,2}} types
See MLton/mlton#334

After converting from one IR to the next, all global data associated
with the previous IR should be dead.  However, it is observed that the
tables used for hash-consing of types seem to be retained after
converting from the corresponding IR.  It is unclear whether this is a
space-safety bug in the compiler or whether the code structure of the
compiler happens to keep those tables live in closures that are
reachable (but unused).

A draft version of this commit used

    val table: t HashSet.t option ref = ref (SOME (HashSet.new {hash = hash}))

    fun lookup (hash, tr) = HashSet.lookup (valOf (!table), hash, ..., ...)

    fun resetTable () = table := NONE

which would have an `Option` exception if the table were actually used
after it was reset; no such exception was observed in a self-compile.