Make `peek.sml` benchmark more realistic
Closes MLton/mlton#312.

Create multiple properties, with multiple types and values, on
multiple property lists.

Previously, with exactly one property with one value created on one
list, `ConstantPropagation` was able to propagate the value through
the program, so that by the end of ssa, the function is a simple loop,
subtracting 1 from one argument, and adding 13 to the other.