Use `staticHeapD` to collect the initial dynamic heap objects
When the mutator marks cards, `staticHeapR` cannot be used for objects
with mutable objptr fields, because the address of an object in
`staticHeapR` won't map to a valid card slot for the write barrier.
Such global objects with mutable objptr fields must be placed in the
dynamic heap and referenced indirectly via a `Global` operand.
However, it is still possible to collect all such objects into a
static heap, which is copied to the initial dynamic heap, rather than
initializing them via the `initGlobals` function.

See https://github.com/MLton/mlton/pull/328#issuecomment-527693766.