Tweak C-side computation of static heap sizes
A "packed" `struct` may still have padding at the end, which is
included in C's `sizeof` calculation.  However, for `memcpy`-ing and
`foreachObjptrInRange`-ing a static heap, we need a byte accurate
size.

We include a sentinel `struct {} end;` field in the static heap
structs to explicitly compute the size of a static heap as
`(pointer)&staticHeap.end - (pointer)&staticHeap`.