Fix latent bug in previous optimization of `SimplifyTypes`
In a previous commit, it was noted that it is impossible for a block
to be called with a value of an uninhabited type.  This was used to
justify not including such a block in the transformed function.  One
example is a block that is the target of a `Case` transfer for a
`Useless` constructor.  The corresponding branch of the `Case`
transfer will be eliminated and the block will not have a use in the
transformed function.

However, it is possible for such a block to be the `cont` or `handle`
of a non-tail call.  While it will be impossible for the called
function to `Return` or `Raise` with a value of an uninhabited type,
the type of the function must be (mostly) preserved.  In particular,
the `returns` and `raises` are transformed via:

    val returns = Option.map (returns, keepSimplifyTypes)
    val raises = Option.map (raises, keepSimplifyTypes)

where `keepSimplifyTypes` drops uninhabited components.  Because the
`returns` and `raises` of the function are not eliminated, the `cont`
and `handle` blocks cannot be eliminated.  On the other hand, when the
block is the target of a `Case` transfer for a `Useless` constructor,
then there will be no use of the block in the transformed program and
the block will be eliminated by `Shrink`.

While it would be possible to adjust the `cont` and `handle` at a
`Call`, it is non-trivial.  For example, a function that purports to
return a value of an uninhabited type cannot actually `Return`, but
could continue to `Raise`.  However, the SSA IR cannot express a
non-tail call with a `handle` but no `cont`, so some dummy `cont`
block would nonetheless be required.  Moreover, it would require
inspecting the `returns` and `raises` types of the called function,
which are not currently recorded in a property for `SimplifyTypes`.
Finally, this kind of calling-convention adjustment is suitably
handled by `RemoveUnused`.