Add `flat-array.4.sml` regression
This regression test constructs a
`(Word14.word * Word10.word * Word8.word) vector`, which is
`DeepFlatten`-ed.
`PackedRepresentation.TupleRep.make: makeSubWord32s` represents the
element as a single `Component.Word`, but while the `Word14` and
`Word10` sub-components are accessed with `Select.IndirectUnpack` (by
reading/writing the whole `Word32` with shifting/masking), the `Word8`
sub-component is accessed with `Select.Indirect` (because the packing
results in the `Word8` sub-component ending up with an 8-bit aligned
offset).

This demonstrates that a version of `RepType.checkOffset` that checks
that the `offset`/`result` exactly corresponds to a component does not
suffice.

Interestingly, this scenario (where there are some non-prim-sized
components combined with some prim-sized components such that a
prim-sized component's offset is properly aligned) never otherwise
arises with the regression tests.

For all prim-sized components, a version of
`PackedRepresentation.TupleRep.make: makeSubWord32sAllPrims` that
represents each element as a separate `Component.Word` and accesses
each element with `Select.Indirect` (as is done for 64-bit and 32-bit
components) is compatible with a version of `RepType.checkOffset` that
checks that the `offset`/`result` exactly corresponds to a
component (because each element is represented as a separate
component).  This scenario does occur often.