Fix infinite loop with jump chaining in `RssaShrink`
MLton/mlton#298 introduced a (simple) jump optimization in
`RssaShrink`.  However, an infinite loop block like

  loop_61 () Jump =
    loop_61 ()

would lead to an infinite loop in `RssaShrink` trying to compute the
replacement label for `loop_61`.

To avoid the infinite loop, we pre-calculate the replacement label for
each block, keeping track of labels seen and marking the looping block
as having no replacement.

Closes MLton/mlton#372