Fix bug in running `blockFn`s for `CML.sync`
When no base event in a CML synchronization is enabled, then all base
events must be blocked.  It is important that all of the `blockFn`s
are executed atomically (along with the prior work of
synchronization), but the previous implementation briefly left the
critical section as it switched from the thread that executed one
`blockFn` to the next.  This would allow the timer interrupt to switch
to another CML thread, without executing all of the `blockFn`s; the
switched-to thread could initiate a synchronization that doesn't
properly see the effect of the switched-from thread's `blockFn`s.

As a concrete example, consider a receiver thread that performs
`CML.select (List.map CML.recvEvt chs)`; that is, it synchronizes on
the choice of receiving from any channel in a list of channels.
Assuming that there is no pending sender, the receiver thread will
execute `blockFn`s that places the receiver thread on each of the
channels.  However, if the timer interrupt arrives during the
execution of the `blockFn`s, then the program can switch to another
thread before the receiver thread is properly placed on each of the
channels.  Now, suppose that the switched-to thread performs `CML.send
ch` for some `ch` in `chs` beyond the point where the `blockFn`s were
successfully executed; that is, it synchronizes on sending to one of
the channels on which the receiver thread has not yet been placed.
Finding no pending receiver, the sender thread executes a `blockFn`
that places the sender thread on the channel.  Control then switches
back to the receiver thread, which continues executing the `blockFn`s,
placing it on the remaining channels, including the channel that has
the blocked sender.  At this point, the program has violated the
invariant that a channel never has both pending senders and pending
receivers and, if there are no other ready threads in the system, will
terminate with an apparent deadlock.

The fix is to simply perform an extra `S.atomicBegin()` immediately
before switching to the temporary thread that will execute the
`blockFn`; see the `ConcurrentMLImplementation` page for details about
why the `blockFn` is executed by a temporary thread.  This ensures
that all of the `blockFn`s are executed in the same critical section
as the prior work of synchronization, without the possibility of
switching to another thread before all `blockFn`s are executed.

Closes MLton/mlton#381.