Prefer jump chaining over block inlining in RssaShrink
Given blocks

  L_A (x, y, z)
    L_B (x, y, z)
  L_B (a, b, c)
    ...

where L_B has exactly one occurrence, there is a choice on how to simplify.  One
option choice is to inline L_B into L_A and the other choice is to replace L_A
with L_B.

Previously, inlining was chosen.  With this commit, replacing is chosen.