Fix bug in RssaShrink
Previously, `RssaShrink` could fail with a `no RssaShrink.replaceVar
property` internal compiler error.  Consider the following:

  L_K (x) Cont =
    L_1 (x)
  L_1 (y) Jump =
    loop (y)
  loop (z) Jump =
    ...
  L_2 (...) Jump =
    ...
    a = y
    ...

L_1 was identified as a candidate for jump chaining and the
`goto L_1 (x)` in L_K was replaced by `goto loop (x)`:

  L_K (x) Cont =
    loop (x)
  loop (z) Jump =
    ...
  L_2 (...) Jump =
    ...
    a = y
    ...

This results in an unbound variable `y`.  (In practice, it led to a
`Fail: y has no RssaShrink.replaceVar property`, because no binding
occurrence of `y` was encountered before the use during the
transformation.)

Now, an occurrence count is calculated for each variable and a block
is a candidate for jump chaining only if all of its arguments have an
occurrence count of 1 (corresponding to their uses in the block's goto
transfer).

Closes MLton/mlton#401.