Avoid defaulting to native codegen on x86-darwin
The C compiler on x86-darwin defaults to producing position
independent code, but the x86 codegen does not properly support
x86-darwin PIC.

x86-darwin barely exists as a platform anymore.  It may be possible to
get the old behavior by compiling with `-codegen 86 -pi-style npi`,
but this is untested.