Fix bug in handling of weak objects during mark-compact GC
During a DFS marking, a weak object should only be linked if its
objptr field is a valid object pointer (and not `BOGUS_OBJPTR`).

This bug was introduced by commit ec2a3afa2, which changed
`is{Objptr,Pointer}` predicates to `is{Objptr,Pointer}InHeap`
predicates.  This was motivated by the introduction of static heaps,
because objects in the static heaps are not traced; in particular,
objects in the immutable static heap are in read-only memory, and
therefore cannot be marked and unmarked.

However, the semantics of these predicates are subtly different.  The
`is{Objptr,Pointer}InHeap` return `TRUE` when the argument is a
non-pointer (i.e., when `is{Objptr,Pointer}` returns `FALSE`).

Commit 2d46cc122 fixed many of the `isPointerInHeap(p)` predicates to
`isPointer(p) and isPointerInHeap(p)`, but missed the `isObjptrInHeap`
that guards the linking of a weak object.