Fix bug in `SimplifyTypes` SSA optimization pass.
19b07c03f attempted to fix a bug in `SimplifyTypes` related to
`MLton_bogus` primitives added by `Useless`.  The essence of the bug
is that is not appropriate to conclude that a tycon has cardinality
`Zero` (e.g., if it is simplified to the point where it has no
constructors) when a "value" of the tycon is created by `MLton_bogus`.

19b07c03f simply forced any tycon used with `MLton_bogus` to have
cardinality `Many`.  However, such a tycon might have all constructors
eliminated and will be marked to be replaced by the `void` type.  In
turn, other types that use the tycon will be retained (with
cardinality `Many`), but an `Option` exception would be raised by
`simplifyUsefulTypes`, which assumes that all components are really
useful (and non-`void`).

This commit adjusts things to force any tycon used with `MLton_bogus`
to have cardinality at least `One`; if such a tycon has all
constructors eliminated, then it will be marked to be replaced by the
`unit` type.

Thanks to Martin Elsman for the bug report.

Closes #419.