Control type-checking of ILs with `-{enable,disable}-pass`
Use `-{enable-disable}-pass '.*:typeCheck'` to control type-checking of ILs in
individual simplify passes, simplify pipelines, and translation passes.

The overall default behavior is the same, but rather than type checking at the
beginning and end of each IL simplify pipeline, we now type check at the end of
each translation pass and the end of each IL simplify pipeline.  (For example,
previously, we type checked at the beginning and end of `ssaSimplify`, but now
we type check at the end of `closureConvert` and the end of `ssaSimplify` (and
not at the beginning of `ssaSimplify`).)

In addition, it is possible to enable type checking for exactly one simplify
pass (e.g., `-enable-pass 'constantPropagation:typeCheck'`) without needing to
type check every pass.