Improve `constToVector` in `ConstantPropagation`
When a `WordXVector` constant is coerced to a `Vector` abstract value,
we should ensure that if the constant becomes `Top`, then the `Vector`
does as well.

In practice, this is never an issue, because `WordXVector` constants
never flow into the same abstract value (and, therefore, such abstract
values are never changed).  The reason is that a `WordXVector`
constant has the type `(word8) vector`; while a `WordXVector` constant
induces a `Const` abstract value for the variable to which it is
initially bound and direct references to that variable fetch the
`Const` abstract value, any participation of this abstract value in a
join must be mediated by flowing through an (initially undefined)
abstract value produced by `Value.fromType` on `(word8) vector`, which
induces a `Vector` abstract value.

This somewhat prohibits the effective constant propagation of
`WordXVector` constants.