Minimize `curSourceSeqsIndex` updates with `-profile time-field`
Previously, `-profile time-field` was implemented by updating the
`curSourceSeqsIndex` field whenever a profile label would be inserted for
`-profile time-label`.  However, profile labels are inserted for code coverage
properties (in particular, at the beginning of every basic block), whether or
not the `sourceSeqsIndex` is changing at the control-flow point.

Now, we only update `curSourceSeqsIndex` at control-flow points where the
`sourceSeqsIndex` changes.  In particular, an update is inserted at the
beginning of each function and at the beginning of each `Cont` and `Handler`
blocks.

The overhead of `-profile time-field` is reduced and, in fact, is now lower than
that of `-profile time-label` (possibly due to the overhead of binary searching
for the label).

$ ./build/bin/mlton -codegen amd64 -profile time-label hamlet.mlb
$ ./hamlet 100 > /dev/null
mlprof -thresh 3.0 hamlet mlmon.out
12.84 seconds of CPU time (2.41 seconds GC)
              function                cur
------------------------------------ -----
<gc>                                 15.8%
_res_Integer.scan                    15.8%
_res_BinaryMapFn.find.mem            10.0%
Lab.compare                           7.7%
_res_BinaryMapFn.insert               6.8%
_res_Sequence.make2                   4.4%
_res_BinaryMapFn.N                    3.5%
_res_PrimSequence.Slice.collate.loop  3.2%
_res_BinaryMapFn.T'                   3.2%

$ ./build/bin/mlton -codegen amd64 -profile time-field hamlet.mlb
$ ./hamlet 100 > /dev/null
mlprof -thresh 3.0 hamlet mlmon.out
11.68 seconds of CPU time (2.53 seconds GC)
              function                cur
------------------------------------ -----
<gc>                                 17.8%
_res_Integer.scan                    15.0%
_res_BinaryMapFn.insert               6.4%
_res_PrimSequence.Slice.collate.loop  5.7%
_res_BinaryMapFn.find.mem             4.9%
_res_BinaryMapFn.T'                   4.9%
_res_Sequence.make2                   4.9%
_res_Integer.scan.negate              4.1%
Lab.compare                           3.9%
Primitive.Exn.mkOverflow.fn           3.7%
_res_BinaryMapFn.N                    3.3%