Simplify signal handling
A POSIX-compliant signal handler should only read and write
`volatile sig_atomic_t` objects and call async-safe functions.

This simplifies signal handling, because it removes the invariant that
`signalIsPending && atomicState == 0` iff `limit = 0`.  It also allows
`atomicState` and `limit` to be non-`volatile`.

On the other hand, it slightly increases the overhead of signal handling,
because a signal check cannot be combined with the common `frontier < limit`
limit check (when `bytesRequested < LIMIT_SLOP`); instead, an explicit
`signalIsPending && atomicState == 0` must be performed.  With the default
compile-time options, the signal check is inserted at every loop header and
anywhere a limit check would be inserted.