Further revise checks for switching to signal in `Thread_atomicEnd`
Revise from

    gcState.atomicState--;
    if (0 == gcState.atomicState
        and gcState.signalsInfo.signalIsPending)
      gc;

to

    tmp = gcState.atomicState - 1;
    gcState.atomicState = tmp;
    if (0 == tmp
        and gcState.signalsInfo.signalIsPending)
      gc;

Since `gcState.atomicState` is `volatile`, it may be necessary to reload and
perform an explicit comparison with zero, rather than using the `isZero` flag of
the status register from the `gcState.atomicState--`.  However, since a signal
handler only ever reads `gcState.atomicState` (and never writes it), the value
of that reload will always be equal to `gcState.atomicState - 1`.