Use 32-bit `Bool_t` rather than 8-bit `bool` for `_import`-ed GC fns
There was a mismatch between the C types and SML types for a number of
`_import`-ed functions that read and write boolean fields of the
`struct GC_state`.  In rare circumstances, this would lead to
erroneous behavior due to mismatch in the calling convention (e.g.,
https://github.com/MLton/mlton/runs/4538320144?check_suite_focus=true).

For example, gcc 9.30 and clang 13.0 compile

    bool GC_getAmOriginal (GC_state s_ }
      return s->amOriginal;
    }

to

    movzbl 0x31(%rdi),%eax
    ret

Although `bool` is 8-bit, the result is nonetheless zero extended to
%eax when returning.  However, clang 11.0 compiles the above function
to

    movb 0x31(%rdi),%al
    ret

In either case, the SML code assumes that a 32-bit value is being
returned and compares the value with 0.  In many scenarios, this was
masked because the native codegen always sets %eax to an
unsigned (typically 0) value when calling a C function (see
https://github.com/MLton/mlton/blob/4a86efc5ed82abeb183cb999bd955339caa8d53a/mlton/codegen/amd64-codegen/amd64-generate-transfers.fun#L1325)
and presumably for the C and llvm codegens, it is often the case that
%eax has zero in the upper bits.  However, if the high bits of %eax
are not zero at the call of `GC_getAmOriginal`, then a `false` result
will erroneously be interpreted as `true` (and, in the case of the
`world*` regression tests, leads to fork-bomb behavior).

This commit simply uses the (32-bit) `Bool_t` in the C code to
establish the expected calling convention.