Include generated `mllex` and `mlyacc` files for bootstrapping
If there is no available `mllex` or `mlyacc` (e.g., when bootstrapping
with another SML compiler), use the saved bootstrap files.

This also slightly simplifies the `remote-bootstrap` target, since
only the compiler proper (and not the tools) needs to be made on the
remote system for the initial boot files.