Accept a non-semicolon terminated final expression in the syntax of programs
The Definition specifies the `program` non-terminal as:

  program ::= topdec ';' programOpt | exp ';' programOpt
  programOpt ::=  | program

So, technically, every program should end with a semicolon.

MLton has always deviated slightly from the Definition in that it has always
accepted a non-semicolon terminated final topdec in the syntax of programs.
MLton's grammar for the `program` non-terminal prior to this commit corresponds
to:

  program ::= topdec semiProgramOpt | exp ';' programOpt
  semiProgramOpt :=  | ';' program
  programOpt ::=  | program

which requires every top-level `exp` to be terminated by a semicolon.  This
seems to be an unnecessary restriction.  Indeed, SML/NJ, PolyML, and MLKit all
accept a non-semicolon terminated final expression in the syntax of
programs.

With this commit, MLton's grammar for the `program` non-terminal corresponds to:

  program ::= topdec semiProgramOpt | exp ';' semiProgramOpt
  semiProgramOpt :=  | ';' program

Recall that the `topdec` non-terminal may be empty.