Properly handle single stack frame in `GC_handleSigProf`
For `-profile time`, when a profiling signal arrives, the `GC_handleSigProf`
handler must determine whether the program is executing in the GC, in mutator
SML code, or in mutator (`_import`-ed) C code.  Whether or not the program is
executing in the GC is determined by a `amInGC` flag in the `struct GC_state`.
Whether the mutator is executing in SML code or C code is determined by mapping
the "return address" of the top-most stack frame to a `struct GC_frameInfo` and
consulting the `const GC_frameKind kind;`.

However, if there is a single frame on the stack, there is no "return address"
of the top-most stack frame.  Rather,
`*((GC_returnAddress*)(s->stackTop - GC_RETURNADDRESS_SIZE))` ends up reading
the `size_t used;` field of `struct GC_stack`.  A subsequent segfault would be
the most likely outcome.

A simple `s->stackTop == s->stackBottom` detects that there is a single frame on
the stack, which must necessarily correspond to the mutator executing SML code.

Although this bug (handling a profiling signal when the mutator is executing
with a single frame on the stack) has been latent for quite some time, it
typically would not lead to failures because, until fcbf9177f, there was a
non-tail call from the RSSA `initGlobals` function to the translated SSA2 `main`
function.  The `initGlobals` executed so quickly that it was very unlikely that
a profiling signal would arrive during its execution; therefore, the main
computation would proceed in a second SML frame on the stack, always having the
`initGlobals` stack frame (and a valid "return address") below it.

With fcbf9177f, the RSSA `initGlobals` function tail calls the translated SSA2
`main` function, making it significantly more likely for the profiling signal to
arrive when there is a single frame on the stack.

Closes MLton/mlton#473.