Sign- or zero-extend 8- and 16-bit args for C calls in x86 and amd64 codegens
The x86 and x86-64 calling conventions require 8- and 16-bit arguments to be
passed as 32-bit values (on the stack for x86 and in registers for x86-64).

The x86-64 SysV ABI doesn't specify the contents of the upper bits for such
argument extensions.  Both gcc and clang sign- or zero-extend according to the
declared type of the argument.  Furthermore, clang assumes such a type-directed
extension of incoming arguments (and can optimize accordingly).  See
https://stackoverflow.com/questions/36706721/is-a-sign-or-zero-extension-required-when-adding-a-32bit-offset-to-a-pointer-for/36760539#36760539.

For example, consider the function:

    unsigned int s8_to_u32(signed char x) {
        return (unsigned int)x;
    }

clang will optimize to

    movl %edi, %eax

while gcc will optimize to

    movsbl %dil, %eax

For slightly increased clang compatibility, the x86 and amd64 codegens are
modified to emit either a signed or zero extension conditional on the declared
type of the argument in the C function's prototype (rather than unconditionally
using a zero extension).