Fix bug in X86GenerateTransfers introduced by 286a54cc0.
Commit 286a54cc0 introduced a dependency on the declared prototype of
a C function when generating the assembly for a C function call, in
order to properly sign extend 8-bit and 16-bit arguments.

Early in the x86 codegen, Word64 values are converted into pairs of
Word32 values, so a C function call with Word64 values will converted
to a call with pairs of Word32 values for those arguments (increasing
the number of arguments relative to the earlier IRs).  But, no such
translation was made to the C function's prototype, which continued to
use Int64 and Word64 types.  This lead to a `Fail: fold2`
internal-compiler error due to the mismatch in the number of actual
arguments and the number of prototype arguments.

A simple solution is to expand the Int64 and Word64 prototype
arguments to pairs of Word32 arguments (signedness does not matter).